feat: refactor get_payment_addr function
- Moved get_payment_addr function to common.py for reuse
- Updated test files to use the common get_payment_addr function
- Removed redundant get_payment_addr implementations from test files
- Ensured caching and funding logic is consistent across tests
