+++ cardano_node_tests/tests/tests_conway/test_pparam_update.py
@@ -1388,23 +1388,12 @@ class TestLegacyProposals:
         cluster: clusterlib.ClusterLib,
     ) -> clusterlib.AddressRecord:
         """Create new payment address."""
-        with cluster_manager.cache_fixture() as fixture_cache:
-            if fixture_cache.value:
-                return fixture_cache.value  # type: ignore
-
-            addr = clusterlib_utils.create_payment_addr_records(
-                f"addr_test_conway_update_proposal_ci{cluster_manager.cluster_instance_num}_0",
-                cluster_obj=cluster,
-            )[0]
-            fixture_cache.value = addr
-
-        # Fund source addresses
-        clusterlib_utils.fund_from_faucet(
-            addr,
+        addr = common.get_payment_addr(
+            name_template=common.get_test_id(cluster),
+            cluster_manager=cluster_manager,
             cluster_obj=cluster,
-            all_faucets=cluster_manager.cache.addrs_data,
+            caching_key=helpers.get_current_line_str(),
         )
-
         return addr
 
     @allure.link(helpers.get_vcs_link())
