+++ cardano_node_tests/tests/tests_conway/test_treasury_withdrawals.py
@@ -726,25 +726,13 @@ class TestMIRCerts:
         cluster: clusterlib.ClusterLib,
     ) -> clusterlib.AddressRecord:
         """Create new payment address."""
-        test_id = common.get_test_id(cluster)
-        with cluster_manager.cache_fixture() as fixture_cache:
-            if fixture_cache.value:
-                return fixture_cache.value  # type: ignore
-
-            addr = clusterlib_utils.create_payment_addr_records(
-                f"{test_id}_payment_addr_0",
-                cluster_obj=cluster,
-            )[0]
-            fixture_cache.value = addr
-
-        # Fund source addresses
-        clusterlib_utils.fund_from_faucet(
-            addr,
+        addr = common.get_payment_addr(
+            name_template=common.get_test_id(cluster),
+            cluster_manager=cluster_manager,
             cluster_obj=cluster,
-            all_faucets=cluster_manager.cache.addrs_data,
+            caching_key=helpers.get_current_line_str(),
             amount=4_000_000,
         )
-
         return addr
 
     @allure.link(helpers.get_vcs_link())
